Check if Bluetooth is enabled prior to scanning
If the user starts the application without BT enabled then a prompt is
shown asking if they would like to turn BT on. Currently if the user has
BT enabled then scanning will commence however this is not true if BT is
turned off at the start. In future their will be a button to start the
scanning process.