Fix issue with recycleview blinking
The issue involves the recycleview updating existing cards and getting
them mixed up. This is because by default the recycleview does not use
use stable ids to prevent this. However, now the recycleview uses stable
ids with the id originating from the hash code of the object within the
sensors adapter.