Favourite button state is now persistent
The button state for each device found is now persistent with the being
displayed in its checked state if the device was favourited in a prior
scan. This is achieved using shared preferences whereby a key (MAC
address) and boolean (true) is stored and recalled when required. If
device has been favourited and then unfavourited the key is deleted. Due
to the use of the apply fn there maybe situations where device state is
not updated instantly. This will not be fixed.