Experiment planner show alert message
The experiment planner is now complete for the time being. It can ask
the user information about how the following experiment shall be
performed and stores this inside a manifest. The manifest is capable of
holding all of the parameters of an experiment and can communicate this
information back to the user. The manifest is expected to be handed over
to the experiment activity when ready.